breaks out tables into their own comps
